requirements.txt
pytest==8.3.2
selene==2.0.0rc9
allure-python-commons==2.13.5
allure-pytest==2.13.5
python-dotenv==1.0.1